requirements.txt
numpy
pandas
scipy
matplotlib
tqdm
product
opt_einsum
pre-commit
black
isort
hydra-core
pathlib
